# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Para servir y administrar restaurantes\nhttps://devsstore.net\n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.2.0",
    "title": "MenuXD",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "http://104.248.75.235:5000",
  "basePath": "/api/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/orders/call/{clientId}/waiter": {
      "get": {
        "description": "Solicita al mozo.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client id",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/": {
      "post": {
        "description": "Crea una nueva orden.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Creado",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "put": {
        "description": "Actualiza una Orden por su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID para actualizar una orden.",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/add/{id}/client/{clientId}": {
      "put": {
        "description": "Actualiza una Orden por su id agregando nuevos items.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la orden.",
            "required": true,
            "type": "number"
          },
          {
            "name": "clientId",
            "in": "path",
            "description": "ID del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Item"
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/client/{clientId}": {
      "get": {
        "description": "Retorna un arreglo de ordenes por su client id",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "ID del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/orders/active/{clientId}": {
      "get": {
        "description": "Retorna un arreglo de ordenes activas por su client id.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "ID del cliente.",
            "required": true,
            "type": "number"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Error en la solicitud"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Item": {
      "type": "object",
      "required": [
        "order_id"
      ],
      "properties": {
        "id": {
          "type": "number",
          "example": 1.0
        },
        "order_id": {
          "type": "number",
          "example": 1.0
        },
        "dish": {
          "type": "object",
          "properties": {}
        },
        "mount": {
          "type": "number",
          "example": 2.0
        },
        "takeaway": {
          "type": "boolean",
          "example": false
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2.0
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        },
        "canceled": {
          "type": "boolean",
          "example": true
        },
        "table": {
          "type": "object",
          "properties": {}
        },
        "client_id": {
          "type": "number",
          "example": 1.0
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
  }
}